torch tests renamed
